ER_MAP/envs/openenv_triage/server-requirements.txt
# Container-side requirements for the ER-MAP TriageOpenEnv server.
#
# Kept separate from the repo-root requirements.txt so HF Space rebuilds
# only depend on the wrapper's needs (not the full RL training stack).
openenv-core>=0.2.3
gymnasium>=0.29.0
groq>=0.18.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
requests>=2.25.0
pyyaml>=6.0
